improved storage flow
