+++ examples/rust/4-storage/README.md
@@ -1,40 +1,31 @@
-# Storage
+# Storage Lifecycle
 
-Read and write data on Pubky homeservers.
+Write, read back, and delete a file on your own Pubky homeserver.
 
-## Public Read (unauthenticated)
+This example defaults to:
 
-Get public data from any user's homeserver storage — no keys needed.
+- path: `/pub/my-app/data.json`
+- content: `my data`
+- recovery file: `../sample_recovery.key`, which has an empty passphrase
 
-```bash
-cargo run --bin storage read pubky<user>/pub/<path>
+If the recovery file cannot be decrypted with an empty passphrase, the CLI prompts for a passphrase.
 
-# example
-cargo run --bin storage read pubky://operrr8wsbpr3ue9d4qj41ge1kcc6r7fdiy6o3ugjrrhi4y77rdo/pub/pubky.app/posts/0033X02JAN0SG
+## Usage
 
-# testnet mode
-cargo run --bin storage read -- --testnet pubky<user>/pub/<path>
-```
-
-## Authenticated Write
-
-Write, read back, and delete a file on your own homeserver.
-
-### Recovery File
-
-This example defaults to `../sample_recovery.key`, which has an empty passphrase.
-If that sample key cannot be decrypted with an empty passphrase, the CLI prompts for a passphrase.
-
-Sign up to a homeserver first using the [signup example](../2-signup).
+Run against a local testnet:
 
 ```bash
-cargo run --bin storage write
-
-# custom path and content
-cargo run --bin storage write --content "my data" /pub/my-app/data.json
-
-# with a custom recovery file
-cargo run --bin storage write </path/to/recovery file>
+cargo run --bin storage -- --testnet
 ```
 
-You can pass `--testnet` to use the local testnet defaults.
+Use a custom path or content:
+
+```bash
+cargo run --bin storage -- --testnet /pub/my-app/data.json --content "my data"
+```
+
+Use a custom recovery file:
+
+```bash
+cargo run --bin storage -- --testnet --recovery-file </path/to/recovery file>
+```
